+++ torchmetrics/functional/text/bleu.py
@@ -183,7 +183,7 @@ def bleu_score(
         and Skip-Bigram Statistics by Chin-Yew Lin and Franz Josef Och `Machine Translation Evolution`_
     """
     warnings.warn(
-        "Input order of preds and targets were changed to target firsts and predictions \
+        "Input order of targets and preds were changed to predictions firsts and targets \
                    second in v0.7. Warning will be removed in v0.8"
     )